Refactor galaxy_install install managers into galaxy.
