+++ lib/galaxy/tool_shed/galaxy_install/install_manager.py
