+++ lib/galaxy/tool_shed/galaxy_install/installed_repository_manager.py
