+++ lib/tool_shed/galaxy_install/repair_repository_manager.py
@@ -1,6 +1,7 @@
 import logging
 import tempfile
 
+from galaxy.tool_shed.galaxy_install import install_manager
 from galaxy.tool_shed.galaxy_install.repository_dependencies import repository_dependency_manager
 from galaxy.tool_shed.galaxy_install.tools import tool_panel_manager
 from galaxy.tool_shed.util import basic_util
@@ -11,7 +12,7 @@ from galaxy.tool_shed.util import tool_dependency_util
 from galaxy.tool_shed.util.container_util import get_components_from_key
 from galaxy.util import unicodify
 from galaxy.util.tool_shed import common_util
-from . import install_manager
+
 
 log = logging.getLogger(__name__)
